This project is based on breast cancer diagnostic dataset of various patients from Kaggle for predictive analysis. It is a classification problem and our goal is to predict the type of tumor whether it is benign or malignant based on other given data.  Here, Random Forest Algorithm and Artificial Neural Network (ANN) are used in R for the predictions.
This project is based on breast cancer diagnostic dataset of various patients from Kaggle for predictive analysis.
It is a classification problem and our goal is to predict the type of tumor whether it is benign or malignant based on other given data.
Here, Random Forest Algorithm and Artificial Neural Network (ANN) are used in R for the predictions.